+++ libraries/base/Data/Ord.hs
@@ -7,7 +7,7 @@
 -- Module      :  Data.Ord
 -- Copyright   :  (c) The University of Glasgow 2005
 -- License     :  BSD-style (see the file libraries/base/LICENSE)
--- 
+--
 -- Maintainer  :  libraries@haskell.org
 -- Stability   :  stable
 -- Portability :  portable
@@ -23,12 +23,18 @@ module Data.Ord (
    comparing,
  ) where
 
+import Data.Bits (Bits, FiniteBits)
+import Foreign.Storable (Storable)
+import GHC.Arr (Ix)
 import GHC.Base
-import GHC.Show
-import GHC.Read
+import GHC.Enum (Bounded, Enum)
+import GHC.Float (Floating, RealFloat)
 import GHC.Num
+import GHC.Read
+import GHC.Real (Fractional, Integral, Real, RealFrac)
+import GHC.Show
 
--- | 
+-- |
 -- > comparing p x y = compare (p x) (p y)
 --
 -- Useful combinator for use in conjunction with the @xxxBy@ family
@@ -54,6 +60,18 @@ newtype Down a = Down
       , Num       -- ^ @since 4.11.0.0
       , Semigroup -- ^ @since 4.11.0.0
       , Monoid    -- ^ @since 4.11.0.0
+      , Bits       -- ^ @since 4.14.0.0
+      , Bounded    -- ^ @since 4.14.0.0
+      , Enum       -- ^ @since 4.14.0.0
+      , FiniteBits -- ^ @since 4.14.0.0
+      , Floating   -- ^ @since 4.14.0.0
+      , Fractional -- ^ @since 4.14.0.0
+      , Integral   -- ^ @since 4.14.0.0
+      , Ix         -- ^ @since 4.14.0.0
+      , Real       -- ^ @since 4.14.0.0
+      , RealFrac   -- ^ @since 4.14.0.0
+      , RealFloat  -- ^ @since 4.14.0.0
+      , Storable   -- ^ @since 4.14.0.0
       )
 
 -- | This instance would be equivalent to the derived instances of the